Refactor for future reusability.
