+++ src/include/OSL/rendererservices.h
@@ -7,6 +7,7 @@
 
 #include <OSL/encodedtypes.h>
 #include <OSL/oslconfig.h>
+#include <OSL/variant.h>
 
 
 OSL_NAMESPACE_ENTER
@@ -37,291 +38,13 @@ enum class AttributeSpecBuiltinArg {
     Derivatives,           // bool
     Type,                  // TypeDesc_pod
     ArrayIndex,            // int, Always zero for non-indexed array lookups.
-    ArrayLookup,           // bool
+    IsArrayLookup,           // bool
     ObjectName,            // const char* (TODO: change to ustringhash)
     AttributeName,         // const char* (TODO: change to ustringhash)
 };
 
-// TODO: When minimum required C++ reaches C++17, replace with std::variant.
-class AttributeSpecArg {
-public:
-    enum class Type {
-        Unspecified = 0,
-        Builtin,
-        Bool,
-        Int8,
-        Int16,
-        Int32,
-        Int64,
-        UInt8,
-        UInt16,
-        UInt32,
-        UInt64,
-        Float,
-        Double,
-        Pointer,
-        UString,
-        UStringHash,
-    };
-
-private:
-    union {
-        AttributeSpecBuiltinArg m_builtin;
-        bool m_bool;
-        int8_t m_int8;
-        int16_t m_int16;
-        int32_t m_int32;
-        int64_t m_int64;
-        uint8_t m_uint8;
-        uint16_t m_uint16;
-        uint32_t m_uint32;
-        uint64_t m_uint64;
-        float m_float;
-        double m_double;
-        void* m_ptr;
-        ustring m_ustring;
-        ustringhash m_ustringhash;
-    };
-    Type m_type;
-
-public:
-    AttributeSpecArg() : m_type(Type::Unspecified) {}
-    AttributeSpecArg(AttributeSpecBuiltinArg arg)
-        : m_builtin(arg), m_type(Type::Builtin)
-    {
-    }
-    AttributeSpecArg(bool arg) : m_bool(arg), m_type(Type::Bool) {}
-    AttributeSpecArg(int8_t arg) : m_int8(arg), m_type(Type::Int8) {}
-    AttributeSpecArg(int16_t arg) : m_int16(arg), m_type(Type::Int16) {}
-    AttributeSpecArg(int32_t arg) : m_int32(arg), m_type(Type::Int32) {}
-    AttributeSpecArg(int64_t arg) : m_int64(arg), m_type(Type::Int64) {}
-    AttributeSpecArg(uint8_t arg) : m_uint8(arg), m_type(Type::UInt8) {}
-    AttributeSpecArg(uint16_t arg) : m_uint16(arg), m_type(Type::UInt16) {}
-    AttributeSpecArg(uint32_t arg) : m_uint32(arg), m_type(Type::UInt32) {}
-    AttributeSpecArg(uint64_t arg) : m_uint64(arg), m_type(Type::UInt64) {}
-    AttributeSpecArg(float arg) : m_float(arg), m_type(Type::Float) {}
-    AttributeSpecArg(double arg) : m_double(arg), m_type(Type::Double) {}
-    AttributeSpecArg(void* arg) : m_ptr(arg), m_type(Type::Pointer) {}
-    AttributeSpecArg(ustring arg) : m_ustring(arg), m_type(Type::UString) {}
-    AttributeSpecArg(ustringhash arg)
-        : m_ustringhash(arg), m_type(Type::UStringHash)
-    {
-    }
-
-    AttributeSpecArg(const AttributeSpecArg& other)
-    {
-        memcpy((void*)this, &other, sizeof(AttributeSpecArg));
-    }
-    AttributeSpecArg(AttributeSpecArg&& other)
-    {
-        memcpy((void*)this, &other, sizeof(AttributeSpecArg));
-    }
-
-    ~AttributeSpecArg() {}
-
-    AttributeSpecArg& operator=(const AttributeSpecArg& other)
-    {
-        memcpy((void*)this, &other, sizeof(AttributeSpecArg));
-        return *this;
-    }
-    AttributeSpecArg& operator=(AttributeSpecArg&& other)
-    {
-        memcpy((void*)this, &other, sizeof(AttributeSpecArg));
-        return *this;
-    }
-
-    Type type() const { return m_type; }
-
-    template<typename T> bool is_holding() const
-    {
-        if (std::is_same<T, AttributeSpecBuiltinArg>::value)
-            return m_type == Type::Builtin;
-        if (std::is_same<T, bool>::value)
-            return m_type == Type::Bool;
-        if (std::is_same<T, int8_t>::value)
-            return m_type == Type::Int8;
-        if (std::is_same<T, int16_t>::value)
-            return m_type == Type::Int16;
-        if (std::is_same<T, int32_t>::value)
-            return m_type == Type::Int32;
-        if (std::is_same<T, int64_t>::value)
-            return m_type == Type::Int64;
-        if (std::is_same<T, uint8_t>::value)
-            return m_type == Type::UInt8;
-        if (std::is_same<T, uint16_t>::value)
-            return m_type == Type::UInt16;
-        if (std::is_same<T, uint32_t>::value)
-            return m_type == Type::UInt32;
-        if (std::is_same<T, uint64_t>::value)
-            return m_type == Type::UInt64;
-        if (std::is_same<T, float>::value)
-            return m_type == Type::Float;
-        if (std::is_same<T, double>::value)
-            return m_type == Type::Double;
-        if (std::is_same<T, void*>::value)
-            return m_type == Type::Pointer;
-        if (std::is_same<T, ustring>::value)
-            return m_type == Type::UString;
-        if (std::is_same<T, ustringhash>::value)
-            return m_type == Type::UStringHash;
-
-        return false;
-    }
-
-    template<typename T> T get() const
-    {
-        OSL_DASSERT(false);
-        return T();
-    }
-};
-
-
-template<>
-inline AttributeSpecBuiltinArg
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<AttributeSpecBuiltinArg>());
-    return m_builtin;
-}
-
-template<>
-inline bool
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<bool>());
-    return m_bool;
-}
-
-template<>
-inline int8_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<int8_t>());
-    return m_int8;
-}
-
-template<>
-inline int16_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<int16_t>());
-    return m_int16;
-}
-
-template<>
-inline int32_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<int32_t>());
-    return m_int32;
-}
-
-template<>
-inline int64_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<int64_t>());
-    return m_int64;
-}
-
-template<>
-inline uint8_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<uint8_t>());
-    return m_uint8;
-}
-
-template<>
-inline uint16_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<uint16_t>());
-    return m_uint16;
-}
-
-template<>
-inline uint32_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<uint32_t>());
-    return m_uint32;
-}
-
-template<>
-inline uint64_t
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<uint64_t>());
-    return m_uint64;
-}
-
-template<>
-inline float
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<float>());
-    return m_float;
-}
-
-template<>
-inline double
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<double>());
-    return m_double;
-}
-
-template<>
-inline void*
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<void*>());
-    return m_ptr;
-}
-
-template<>
-inline ustring
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<ustring>());
-    return m_ustring;
-}
-
-template<>
-inline ustringhash
-AttributeSpecArg::get() const
-{
-    OSL_DASSERT(is_holding<ustringhash>());
-    return m_ustringhash;
-}
-
-// The AttributeGetterSpec is never in device code, so it can be a bit more
-// complex and have dynamic allocations.  Although we do want to
-// use an interface and not allow direct data member access
-class AttributeGetterSpec {
-    ustring m_function_name;
-    std::vector<AttributeSpecArg> m_args;
-
-public:
-    template<typename... ArgListT>
-    void set(ustring function_name, ArgListT... args)
-    {
-        m_function_name = function_name;
-        m_args.clear();
-        m_args.insert(m_args.end(), std::initializer_list<AttributeSpecArg> {
-                                        AttributeSpecArg { args }... });
-    }
-
-    const ustring& function_name() const { return m_function_name; }
-    size_t arg_count() const { return m_args.size(); }
-    const AttributeSpecArg& arg(size_t i) const
-    {
-        OSL_DASSERT(i < m_args.size());
-        return m_args[i];
-    }
-};
-
+using AttributeSpecArg    = ArgVariant<AttributeSpecBuiltinArg>;
+using AttributeGetterSpec = FunctionSpec<AttributeSpecArg>;
 
 // Turn off warnings about unused params for this file, since we have lots
 // of declarations with stub function bodies.
